fix mypy errors; upgrade project management to poetry; WIP move API components to own router modules for better readability
