+++ mypy.ini
@@ -4,3 +4,6 @@
 warn_return_any = False
 warn_unused_configs = True
 warn_incomplete_stub = False
+namespace_packages = True
+ignore_missing_imports = True
+check-untyped-defs = True
