--- requirements.txt
@@ -1,11 +1,0 @@
-aars~=0.5.8
-aiohttp
-aleph-sdk-python==0.6.0
-fastapi
-pandas
-pydantic
-semver
-base58
-pynacl
-numpy
-requests
